Move lightning address decode to general utils
