+++ src/routes/lnurl/utils.ts
@@ -40,19 +40,3 @@ class Signer {
     return toHexString(DER)
   }
 }
-
-const usernameRegex = /^[a-z0-9-_.]+$/
-
-const domainRegex =
-  /^((?!-))(xn--)?[a-z0-9][a-z0-9-_]{0,61}[a-z0-9]{0,1}\.(xn--)?([a-z0-9-]{1,61}|[a-z0-9-]{1,30}\.[a-z]{2,})$/
-
-export function decodeLightningAddress(val: string): { username: string; domain: string } | null {
-  const [username, domain] = val.split('@')
-
-  // check valid username && valid domain
-  if (!usernameRegex.test(username) || !domainRegex.test(domain)) {
-    return null
-  }
-
-  return { username, domain }
-}
